refactor: Disable custom Material color scheme in CodeEditor
This commit disables the application of a custom Material Design-styled color scheme to the `CodeEditor` widget.

By commenting out the call to `editor.setColorScheme(getMaterialStyledColorScheme(editor))`, the editor will now default to its standard color scheme instead of the custom one. This change simplifies the editor's setup logic.

The following file was modified:
- `app/src/main/java/pro/sketchware/utility/EditorUtils.java`